Phase 1 bottleneck resolved with Kociemba's 16-fold symmetries + small optiomisation on hot loop

+++ Assets/Scripts/Solver/IDAStarSolver.cs
@@ -10,10 +10,10 @@ namespace Assets.Scripts.Solver
         public int InitialBound;
         public int FinalBound;
         public int BoundIterations;
-        public int NodesVisited;
-        public int NodesExpanded;
-        public int ChildrenGenerated;
-        public int PrunedByHeuristic;
+        public long NodesVisited;
+        public long NodesExpanded;
+        public long ChildrenGenerated;
+        public long PrunedByHeuristic;
         public int MaxDepthReached;
         public long ElapsedMilliseconds;
     }
